+++ bark/models/behavior/motion_primitives/primitives/primitive_const_acc_change_to_left.cpp
@@ -12,7 +12,7 @@ bark::models::behavior::primitives::PrimitiveConstAccChangeToLeft::
     : BehaviorModel(params),
       PrimitiveConstAccStayLane(params),
       min_length_(params->GetReal(
-          "MinLength", "Minimum length of lane to change to", 0.0f)) {}
+          "MinLength", "Minimum length of lane to change to", 0.0)) {}
 bark::world::LaneCorridorPtr bark::models::behavior::primitives::
     PrimitiveConstAccChangeToLeft::SelectTargetCorridor(
         const bark::world::ObservedWorld& observed_world,
@@ -41,7 +41,7 @@ bool bark::models::behavior::primitives::PrimitiveConstAccChangeToLeft::
     //! lane
     const Point2d point_on_target_line =
         GetNearestPoint(adjacent_corridors.left->GetCenterLine(), ego_pos);
-    float remaining_length =
+    double remaining_length =
         adjacent_corridors.left->LengthUntilEnd(point_on_target_line);
     satisfied = remaining_length >= min_length_;
   }
