+++ bark/models/behavior/motion_primitives/primitives/primitive_const_acc_change_to_left.hpp
@@ -34,7 +34,7 @@ class PrimitiveConstAccChangeToLeft : public PrimitiveConstAccStayLane {
   std::string GetName() const override;
 
  private:
-  float min_length_;
+  double min_length_;
 };
 
 }  // namespace primitives
