+++ bark/models/behavior/rule_based/lane_change_behavior.hpp
@@ -165,7 +165,7 @@ class BehaviorLaneChangeRuleBased : public BehaviorIDMLaneTracking {
     return state[VEL_POSITION];
   }
 
-  virtual Trajectory Plan(float delta_time,
+  virtual Trajectory Plan(double delta_time,
                           const ObservedWorld& observed_world);
 
   virtual std::shared_ptr<BehaviorModel> Clone() const;
