+++ bark/world/evaluation/base_evaluator.hpp
@@ -20,7 +20,7 @@ class World;
 class ObservedWorld;
 namespace evaluation {
 
-typedef boost::variant<float, bool, std::optional<bool>, std::string, int> EvaluationReturn;
+typedef boost::variant<double, bool, std::optional<bool>, std::string, int> EvaluationReturn;
 
 class BaseEvaluator {
  public:
